feat #1 - Initial web repository setup
Move Next.js app files into src/app and replace the default layout/page with a ThemeProvider-based RootLayout and a simple Home page. Add editor/formatter configs (.prettierrc, .prettierignore, .eslintignore), adjust ESLint to allow <img> usage, and update tsconfig paths to use @/ -> src/*. Add project docs (AGENTS.md, CLAUDE.md) outlining stack and guidelines, create gitkeep placeholders for feature folders, and remove unused default public SVGs. Update package.json (scripts and new deps: clsx, next-themes, tailwind-merge, zustand) and regenerate package-lock.json accordingly. README cleaned up to remove duplicated boilerplate/features section.
