+++ .eslintignore
@@ -1,0 +1,6 @@
+.next
+node_modules
+out
+build
+dist
+next-env.d.ts
