+++ .prettierignore
@@ -1,0 +1,7 @@
+.next
+node_modules
+out
+build
+dist
+*.lock
+package-lock.json
