+++ README.md
@@ -70,54 +70,6 @@ fifawcp/
 - `npm run lint` - Run ESLint
 - `npm run format` - Format code with Prettier
 
-## 🎯 Key Features
-
-- **Match Predictions:** Pick scores for all World Cup matches
-- **Live Scoring:** Real-time points calculation
-- **Leaderboards:** Track rankings against other players
-- **Bracket Simulation:** Knockout stage predictions
-- **Time Locking:** Picks locked 5 minutes before kickoff
-- **Dark Mode:** Full theme support
-- **Mobile-First:** Optimized for all devices
-
-## 🏗️ Development Guidelines
-
-### Code Standards
-
-- **TypeScript:** Strict mode, no `any` types
-- **Components:** PascalCase naming
-- **Hooks:** camelCase with `use` prefix
-- **Server Components:** Default to RSC, use `'use client'` only when needed
-- **Styling:** Tailwind utility classes, mobile-first approach
-
-### Import Aliases
-
-Use `@/` for absolute imports from the `src` directory:
-
-```typescript
-import { calculatePoints } from "@/lib/scoring";
-import type { Match, Pick } from "@/types";
-import { useMatchLock } from "@/hooks/useMatchLock";
-```
-
-### Time Handling
-
-All match times are stored in UTC. Convert to local time only in the UI layer.
-
-## 📝 Architecture Decisions
-
-- **Server-First:** Leverage React Server Components for data fetching
-- **Type Safety:** Centralized type definitions in `src/types`
-- **Scoring Logic:** Deterministic calculation in `src/lib/scoring.ts`
-- **Accessibility:** ARIA labels and keyboard navigation throughout
-
-## 🤝 Contributing
-
-1. Follow the coding standards in `AGENTS.md`
-2. Use conventional commits (e.g., `feat:`, `fix:`, `docs:`)
-3. Run `npm run lint` and `npm run format` before committing
-4. Never commit API keys or sensitive data
-
 ## 📄 License
 
 Private project - All rights reserved
